config: Do not set provider to github for theme
The theme will add a download link to the master.zip, which is 1. not
the version that people should fetch, and 2. not usable because the
submodule is missing.

So disable the gitlab integration. The only regression we have is that
the number of stars is not shown anymore with this.